+++ src/haven/pathfinder/Node.java
@@ -12,13 +12,13 @@ public class Node
     	NORMAL(3),
     	SWAMP(6),
     	THICKET(6),
-    	BLOCK(Long.MAX_VALUE),
-    	BLOCK_DYNAMIC(Long.MAX_VALUE),
+    	BLOCK(Integer.MAX_VALUE),
+    	BLOCK_DYNAMIC(Integer.MAX_VALUE),
     	NOT_IMPLEMENTED(3);
     	
-    	private final long w;
-	    Type(long w) { this.w = w; }
-	    public long getValue() { return w; }
+    	private final int w;
+	    Type(int w) { this.w = w; }
+	    public int getValue() { return w; }
     }
     
     public int x, y;
